pyproject.toml
[project]
name = "bryllup"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas Johan Boesen",email = "lucasjohanboesen123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "qrcode (>=8.2,<9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
